Revise monitoring code; new outputs.
Revision: use matrices to store survey data instead of around 30 separate
vectors (each with its own reporting function and output code).
Outputs are no longer ordered by measure (order of output changes), but this
is should be unimportant (i.e. assuming no tools rely on the order).
Also added some new outputs (far easier to do now).

+++ include/Clinical/Episode.h
@@ -111,7 +111,7 @@ private:
   /// Timestep of event (TIMESTEP_NEVER if no event).
   TimeStep _time;
   /// Survey during which the event occured
-  int _surveyPeriod;
+  size_t _surveyPeriod;
   /// Age group of the individual when the episode's first bout occurred
   Monitoring::AgeGroup _ageGroup;
   /// Descriptor of state, containing reporting info. Not all information will
